requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
cloudinary==1.44.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-parler==2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ffmpeg-python==0.2.0
future==1.0.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-tools==1.71.0
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
multidict==6.4.3
mutagen==1.47.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
qrcode==8.1
redis==6.2.0
reportlab==4.4.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.10.0
